backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic==2.5.3
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Async Tasks
celery==5.3.6
celery[redis]==5.3.6
flower==2.0.1

# Email
sendgrid==6.11.0
jinja2==3.1.3

# ML/Data Science
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
scipy==1.11.4
joblib==1.3.2

# API Clients
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# Utils
python-dateutil==2.8.2
pytz==2023.3.post1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
ipython==8.20.0

# Logging
loguru==0.7.2